feat: create blog data module with posts, categories, and tags
