+++ src/app/blogs/data.ts
@@ -1,0 +1,82 @@
+export const blogPosts = [
+  {
+    id: '1',
+    title: 'Why Blood Donation Is The Most Noble Act Of Kindness',
+    slug: 'why-blood-donation-is-noble',
+    content:
+      'Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam...',
+    excerpt:
+      'Discover why blood donation is considered one of the most impactful ways to help others and save lives...',
+    featuredImage: '/images/blood-donation-hero.jpg',
+    published: true,
+    authorId: 'author1',
+    createdAt: '2024-03-27T10:00:00Z',
+    category: 'DONATION'
+  },
+  {
+    id: '2',
+    title: 'Understanding Blood Types: A Complete Guide',
+    slug: 'blood-types-guide',
+    content:
+      'Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam...',
+    excerpt:
+      'Learn about different blood types, compatibility, and why knowing your blood type is crucial...',
+    featuredImage: '/images/blood-types.jpg',
+    published: true,
+    authorId: 'author2',
+    createdAt: '2024-03-27T09:00:00Z',
+    category: 'EDUCATION'
+  },
+  {
+    id: '3',
+    title: 'Common Myths About Blood Donation Debunked',
+    slug: 'blood-donation-myths',
+    content:
+      'Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam...',
+    excerpt:
+      'We address and debunk common misconceptions about blood donation to help you make informed decisions...',
+    featuredImage: '/images/myths.jpg',
+    published: true,
+    authorId: 'author1',
+    createdAt: '2024-03-27T08:00:00Z',
+    category: 'AWARENESS'
+  }
+]
+
+export const categories = [
+  {
+    id: '1',
+    name: 'Blood Donation',
+    slug: 'blood-donation',
+    count: 340
+  },
+  {
+    id: '2',
+    name: 'Health Tips',
+    slug: 'health-tips',
+    count: 74
+  },
+  {
+    id: '3',
+    name: 'Medical News',
+    slug: 'medical-news',
+    count: 41
+  },
+  {
+    id: '4',
+    name: 'Awareness',
+    slug: 'awareness',
+    count: 35
+  }
+]
+
+export const tags = [
+  'Health',
+  'Blood Types',
+  'Donation',
+  'Medical',
+  'Awareness',
+  'Community',
+  'Healthcare',
+  'Volunteers'
+]
